BIG REFACTOR: WIP removal of painful to maintain AST dumper
These changes are the first proper litmus test for the node metadata
system of the AST. There is now a new piece of data associated with each
AST node which describes all the child nodes that it holds. This can
be used to in a much more generic and simpler way dump the AST.

The really cool thing is I was able to remove the notourious `NODE`
macro from syn in favour of a `start_node`, `end_node` API. This form
allows us to write the parser code and simultaneously describe the
child parent relationships. This makes maintainence much better, I don't
have to keep the dumper in sync with new constructs that I can parse,
this will now be automatic.