Add if statement + expose iota as ffi to python
These chnages both add support for if statements, which now
also raised an ambiguity with struct literal syntax which is
resolved by having `Foo.{ 10 }` instead of `Foo{ 10 }`.

More tests are needed for if statments, however along with these
changes ground work has been put in to reduce the bloat in syntax
tests due to dumping the entire AST for tests that are not testing
the whole AST. This has been done by reworking the build system
to allow for PIC targets to expose a `libiota.so` library which is
now exposed to python through the `ffi/iotalib.py` file.

This already as a nice bonus has increased the speed of the tests
noticeably, but a refactor of the tests will be needed to only
test the subtree that is necessary, this way we can remove the annoying
minmal expression entry points in `t/syn/test_expression.py`, i.e:

```
let _ = x * y;
```