Optimize Arena allocator
Using gprof I noticed that 55% of time was being spent calling the
`try_alloc_in_block` function. For an AST of around 30_000 nodes you
would have ~15 million calls to `try_alloc_in_block`. The first
thing I did was add geometric growth to the block size, so that
when you need to create a new block in the arena you allocate it
double the size of the previously allocated block. This saves a few
million calls.

The massive and main saving however was by separating dynamic arrays
who's ownership is adopted by the arena into their own block list
in the arena. This means that calls to `arena_alloc` don't pay the cost
of skipping these arrays that are fully used, this seems to be the root
cause of the issue. Now calls to `try_alloc_in_block` for the same
number of nodes is now only ~150_000! Just 5 calls per AST node on
average. This is super cheap too so that is really good!