WIP symbol table generation
These changes add an initial symbol table pass in `sem`. Right now it
just inserts struct declarations in the global scope and struct fields in a
sub scope.

Along with these changes I reworked some of the hashmap and added a
stack implementation for the purpose of keeping track of the parent
context when walking the AST. Right now I do not have typed wrappers on
the stack, but maybe I should. Also this is first time I have
implemented a segmented stack and it took a few iterations before it was
not a complete bug farm (might still be :')). The main thing that was
really painful was getting the alignment of the allocated object and the
saved base pointer.

In the end abstracting the implementation made things a lot clearer, I
have a concept of a `Frame` which holds the base pointer of the previous
frame as well as a flexible data member for storing the object and the
offset into this data (based on alignment requirement of object stored
    there). Also having `FrameDescriptor` which basically stores a
bundle pointers: aligned and unaligned varients of the frame and object
pointer. This simple object allows you to pre-calculate all the
alignment bullshit and use the pointers to actually modify the stacks sp
and bp.