Two pass type resolution + interning.
This commit add two pass type resolution to the compiler. This works by
doing bottom up canonicalisation of syntactic type trees. That allows
you to bubble types up to declarations they need to be associated with
(e.g. type declarations, variable declarations and function parameters).

In addition type interning is being done. This requires having a both
a deep comparison and hash function for canonical types. No hashing is
needed after resolution as types are referenced using a rolling id which
is valid for a flat array which stores the actual types.

This flat array could in the future be much easier to serialize into the
binary for runtime type information.